Address inline thinking review feedback and fix build CI
